Changes on the Layout, homepage, services and contact me.
For the home page I added a new button. For the service area I deleted the big black area and shorten the version of the full stack option. For the tech, icons I've re-done the whole grid system and for the contact area, I simpled add my own personal information.